+++ TabletopTweaks-Core/NewComponents/RemoveBuffRankOnAttack.cs
@@ -10,16 +10,16 @@ namespace TabletopTweaks.Core.NewComponents {
     [AllowedOn(typeof(BlueprintBuff))]
     [TypeId("00008ebd4ddb41e6895baa47a7f19708")]
     public class RemoveBuffRankOnAttack : UnitFactComponentDelegate,
-        IInitiatorRulebookHandler<RuleAttackWithWeapon>,
-        IRulebookHandler<RuleAttackWithWeapon>, ISubscriber,
+        IInitiatorRulebookHandler<RuleAttackRoll>,
+        IRulebookHandler<RuleAttackRoll>, ISubscriber,
         IInitiatorRulebookSubscriber {
 
         // Token: 0x0600A2AE RID: 41646 RVA: 0x000036D8 File Offset: 0x000018D8
-        public void OnEventAboutToTrigger(RuleAttackWithWeapon evt) {
+        public void OnEventAboutToTrigger(RuleAttackRoll evt) {
         }
 
         // Token: 0x0600A2AF RID: 41647 RVA: 0x0027A8A4 File Offset: 0x00278AA4
-        public void OnEventDidTrigger(RuleAttackWithWeapon evt) {
+        public void OnEventDidTrigger(RuleAttackRoll evt) {
             using (ContextData<Kingmaker.UnitLogic.Buffs.BuffCollection.RemoveByRank>.RequestIf(true)) {
                 if (Owner != null) {
                     Owner.Buffs.RemoveFact(base.Fact);